requirements.txt
discord.py==2.4.0
motor==3.6.1
python-dotenv==1.0.1